backend/requirements.txt
# Core
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Cache
redis>=5.0,<8.0
django-redis>=5.4,<7.0

# Task queue
celery>=5.4,<6.0
django-celery-beat>=2.6,<3.0
kombu>=5.3,<6.0

# Storage (S3/MinIO)
boto3>=1.34,<2.0
django-storages>=1.14,<2.0

# Image processing
Pillow>=12.0,<13.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Monitoring
django-prometheus>=2.3,<3.0

# WSGI server
gunicorn>=23.0,<26.0

# Logging
python-json-logger>=2.0,<3.0

# Google OAuth
google-auth>=2.20,<3.0
google-auth-oauthlib>=1.2,<2.0

# API documentation
drf-spectacular>=0.27,<1.0

# AI
google-genai>=1.0,<2.0
langchain>=1.2,<2.0
langgraph>=1.1,<2.0
langchain-google-genai>=4.0,<5.0

# CV parsing
pypdf>=5.0,<7.0
python-docx>=1.0,<2.0

# Web scraping
requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0

# LiveKit
livekit-api>=1.0,<2.0

# Telegram bot
python-telegram-bot>=21.0,<23.0

# Env file loading
python-dotenv>=1.0,<2.0

# Dev tools
ruff>=0.5.0
mypy>=1.10
django-stubs>=5.0
djangorestframework-stubs>=3.15

# Pre-commit
pre-commit>=3.7

# Testing
pytest>=8.0,<10.0
pytest-django>=4.8,<5.0
pytest-cov>=6.0,<8.0
factory-boy>=3.3,<4.0
faker>=37.0,<41.0
weasyprint>=68.0,<70.0
